+++ migrations/config/0007_monitor_kinds2.sql
@@ -1,4 +1,0 @@
-ALTER TYPE monitor_kind ADD VALUE IF NOT EXISTS 'mysql';
-ALTER TYPE monitor_kind ADD VALUE IF NOT EXISTS 'mongodb';
-ALTER TYPE monitor_kind ADD VALUE IF NOT EXISTS 'tls';
-ALTER TYPE monitor_kind ADD VALUE IF NOT EXISTS 'push';